kaem: scaffold/read.kaem.
* .gitignore: Update.
* include/linux/x86/syscall.h: Update for M2-Planet.
* lib/linux/brk.c (brk): Remove cast for M2-Planet.
* lib/linux/x86-mes-m2/syscall.c: New file.
* lib/m2/malloc.c: New file.
* lib/string/memset.c (memset): Update for M2-Planet.
* scaffold/read.c: Update for M2-Planet.
* scaffold/kaem.read: New file.

+++ lib/linux/brk.c
@@ -24,5 +24,5 @@
 long
 brk (void *addr)
 {
-  return _sys_call1 (SYS_brk, (long) addr);
+  return _sys_call1 (SYS_brk, addr);
 }
